+++ src/main/java/co/nvqa/common/core/cucumber/glue/ApiEventsSteps.java
@@ -3,11 +3,14 @@ package co.nvqa.common.core.cucumber.glue;
 import co.nvqa.common.core.client.EventClient;
 import co.nvqa.common.core.cucumber.CoreStandardSteps;
 import co.nvqa.common.core.model.event.Event;
+import co.nvqa.common.core.model.event.EventDetail;
 import co.nvqa.common.core.model.event.Events;
 import io.cucumber.guice.ScenarioScoped;
 import io.cucumber.java.en.Then;
 import java.util.List;
 import java.util.Map;
+import java.util.Objects;
+import java.util.stream.Collectors;
 import javax.inject.Inject;
 import lombok.Getter;
 import org.assertj.core.api.Assertions;
@@ -83,18 +86,50 @@ public class ApiEventsSteps extends CoreStandardSteps {
     final long orderId = Long.parseLong(resolveValue(createdOrderId));
 
     doWithRetry(() -> {
-      final Events actualOrderEvents = getEventClient().getOrderEventsByOrderId(orderId);
+      final List<Event> actualOrderEvents = getEventClient().getOrderEventsByOrderId(orderId)
+          .getData().stream().filter(e -> Objects.equals(e.getType(), expectedOrderEvent.getType()))
+          .collect(Collectors.toList());
 
-      Assertions.assertThat(actualOrderEvents.getData())
-          .withFailMessage(f("Order events should not empty, order id: %d, event: %s", orderId,
-              expectedOrderEvent)).isNotEmpty();
-
-      Assertions.assertThat(actualOrderEvents.getData()).withFailMessage(
-              f("%s event is NOT published for order id: %s", expectedOrderEvent.getType(), orderId))
-          .anySatisfy(
-              event -> Assertions.assertThat(event.getType())
-                  .isEqualTo(expectedOrderEvent.getType()));
+      Assertions.assertThat(actualOrderEvents).anySatisfy(
+          event -> Assertions.assertThat(event.getType()).withFailMessage(
+              f("Event %s is NOT published for order id: %d", expectedOrderEvent.getType(),
+                  orderId)).isEqualTo(expectedOrderEvent.getType()));
 
     }, String.format("%s event is published for order id %d", expectedOrderEvent, orderId));
+  }
+
+
+  /**
+   * And API Core - Operator verify that event is published with correct details:<br>
+   * | orderId   | {KEY_LIST_OF_CREATED_ORDERS[1].id} |  <br>
+   * | eventType | HUB_INBOUND_SCAN | <br>
+   * | eventData | {"weight":{"old_value":5,"new_value":5},"length":{"new_value":30},"width":{"new_value":10},"height":{"new_value":20}} |<br>
+   **/
+  @Then("API Core - Operator verify that event is published with correct details:")
+  public void operatorVerifiesOrderEventDetails(Map<String, String> dataTableRaw) {
+    Map<String, String> resolvedData = resolveKeyValues(dataTableRaw);
+    final long orderId = Long.parseLong(resolvedData.get("orderId"));
+    final String expectedEventType = resolvedData.get("eventType");
+    final String expectedEventData = resolvedData.get("eventData");
+    final EventDetail expectedEventDetail = fromJsonSnakeCase(expectedEventData, EventDetail.class);
+
+    doWithRetry(() -> {
+      final List<Event> actualOrderEvents = getEventClient().getOrderEventsByOrderId(orderId)
+          .getData().stream().filter(e -> Objects.equals(e.getType(), expectedEventType))
+          .collect(Collectors.toList());
+
+      Assertions.assertThat(actualOrderEvents).anySatisfy(
+          event -> Assertions.assertThat(event.getType()).withFailMessage(
+                  f("Event %s is NOT published for order id: %d", expectedEventType, orderId))
+              .isEqualTo(expectedEventType));
+
+      Assertions.assertThat(actualOrderEvents).anySatisfy(
+          event -> Assertions.assertThat(event.getData()).withFailMessage(
+                  f("Actual data:\n %s \ndoes not match expected data:\n %s", toJson(event.getData()),
+                      toJson(expectedEventDetail))).usingRecursiveComparison()
+              .ignoringExpectedNullFields().isEqualTo(expectedEventDetail));
+
+    }, String.format("%s event is published for order id %d with expected details %s",
+        expectedEventType, orderId, toJson(expectedEventDetail)));
   }
 }