+++ src/main/java/co/nvqa/common/core/cucumber/glue/ApiOrderSteps.java
@@ -46,7 +46,8 @@ public class ApiOrderSteps extends CoreStandardSteps {
    * order with the same tracking id. <br/><br/><b>Note</b>: becareful that you may face unintended
    * order status due to event propagation delay to Core service
    *
-   * @param tracking key that contains order's tracking id, example: KEY_LIST_OF_CREATED_TRACKING_IDS
+   * @param tracking key that contains order's tracking id, example:
+   *                 KEY_LIST_OF_CREATED_TRACKING_IDS
    */
   @When("API Core - Operator get order details for tracking order {string}")
   public void apiOperatorGetOrderDetailsForTrackingOrder(String tracking) {
@@ -71,7 +72,8 @@ public class ApiOrderSteps extends CoreStandardSteps {
    * previous order with the same tracking id. this is intended to check if you have done certain
    * action to same order and you need previous data prior to the action being done
    *
-   * @param tracking key that contains order's tracking id, example: KEY_LIST_OF_CREATED_TRACKING_IDS
+   * @param tracking key that contains order's tracking id, example:
+   *                 KEY_LIST_OF_CREATED_TRACKING_IDS
    */
   @When("API Core - Operator get order details for previous order {string}")
   public void apiOperatorGetOrderDetailsForPreviousOrder(String tracking) {
@@ -107,7 +109,7 @@ public class ApiOrderSteps extends CoreStandardSteps {
         Assertions.assertThat(actualGranularStatus).as("order granular status doesnt match")
             .isEqualToIgnoringCase(expectedGranularStatus);
         return tempOrder;
-      }, "Order client search order by tracking id", 20_000, 3);
+      }, "Order client search order by tracking id", 30_000, 3);
 
       putInList(KEY_LIST_OF_CREATED_ORDERS, order,
           (o1, o2) -> StringUtils.equalsAnyIgnoreCase(o1.getTrackingId(), o2.getTrackingId()));
@@ -277,7 +279,8 @@ public class ApiOrderSteps extends CoreStandardSteps {
     Assertions.assertThat(orderId).as("Order ID").isNotNull();
     String dimensionsJson = data.get("dimensions");
     Assertions.assertThat(dimensionsJson).as("Dimensions").isNotNull();
-    getOrderClient().updateParcelDimensions(Long.parseLong(orderId), new Dimension(dimensionsJson));
+    doWithRetry(() -> getOrderClient().updateParcelDimensions(Long.parseLong(orderId),
+        new Dimension(dimensionsJson)), "Update Order Dimensions");
   }
 
   @When("API Core - Operator bulk tags parcel with below tag:")
@@ -286,6 +289,21 @@ public class ApiOrderSteps extends CoreStandardSteps {
     final Long orderId = Long.valueOf(dataTable.get("orderId"));
     Long tag = Long.valueOf(dataTable.get("orderTag"));
     List<Long> tags = Collections.singletonList(tag);
-    getOrderClient().addOrderLevelTags(orderId, tags);
+    doWithRetry(() -> getOrderClient().addOrderLevelTags(orderId, tags),
+        "Operator bulk tags parcel with below tag");
+  }
+
+
+  /**
+   * @param dataTableRaw <br/> <b>orderId</b>: {KEY_LIST_OF_CREATED_ORDERS[1].id} <br>
+   *                     <b>newCodAmount</b>: 100
+   */
+  @When("API Core - Operator add/update order COD amount:")
+  public void apiCoreUpdateOrderCodAmount(Map<String, String> dataTableRaw) {
+    final Map<String, String> dataTable = resolveKeyValues(dataTableRaw);
+    final Long orderId = Long.valueOf(dataTable.get("orderId"));
+    final Double newCodAmount = Double.valueOf(dataTable.get("newCodAmount"));
+    doWithRetry(() -> getOrderClient().updateOrderCod(orderId, newCodAmount),
+        "Operator add/update order COD amount:");
   }
 }