+++ src/main/java/co/nvqa/common/core/cucumber/glue/ApiRouteSteps.java
@@ -13,6 +13,8 @@ import co.nvqa.common.core.model.route.BulkAddPickupJobToRouteRequest;
 import co.nvqa.common.core.model.route.BulkAddPickupJobToRouteResponse;
 import co.nvqa.common.core.model.route.MergeWaypointsResponse;
 import co.nvqa.common.core.model.route.MergeWaypointsResponse.Data;
+import co.nvqa.common.core.model.route.ParcelRouteTransferRequest;
+import co.nvqa.common.core.model.route.ParcelRouteTransferResponse;
 import co.nvqa.common.core.model.route.RouteRequest;
 import co.nvqa.common.core.model.route.RouteResponse;
 import co.nvqa.common.core.model.waypoint.Waypoint;
@@ -519,6 +521,34 @@ public class ApiRouteSteps extends CoreStandardSteps {
     doWithRetry(
         () -> getRouteClient().createCoverage(request),
         "create new coverage");
-    put(KEY_COVERAGE,response.getData());
+    put(KEY_COVERAGE, response.getData());
+  }
+
+  /**
+   * Sample:
+   * <p>
+   * API Core - Operator parcel transfer to a new route:
+   * | request | {{"route_id":null,"route_date":"2021-01-19 08:25:13","from_driver_id":null,"to_driver_id":2679,"to_driver_hub_id":3,"orders":[{"tracking_id":"NVSGDIMMI000238068","inbound_type":"VAN_FROM_NINJAVAN","hub_id":3}]|
+   * @param dataTableAsMap Map of data from feature file.
+   */
+  @Given("API Core - Operator parcel transfer to a new route:")
+  public void apiOperatorParcelTransfer(Map<String, String> dataTableAsMap) {
+    Map<String, String> resolvedDataTable = resolveKeyValues(dataTableAsMap);
+
+    ZonedDateTime routeDate = CoreTestUtils.getDateForToday();
+    String formattedRouteDate = DTF_NORMAL_DATETIME.format(
+        routeDate.withZoneSameInstant(ZoneId.of("UTC")));
+
+    ParcelRouteTransferRequest request = fromJsonSnakeCase(resolvedDataTable.get("request"),
+        ParcelRouteTransferRequest.class);
+    if (request.getRouteDate() == null) {
+      request.setRouteDate(formattedRouteDate);
+    }
+
+    doWithRetry(() -> {
+      final ParcelRouteTransferResponse createRouteResponse = getRouteClient()
+          .parcelRouteTransfer(request);
+      put(KEY_LIST_OF_CREATED_ROUTES, createRouteResponse.getRoutes());
+    }, "parcel route transfer");
   }
 }